pyproject.toml
[tool.poetry]
name = "westfarmers-case-study"
version = "0.1.0"
description = "Example case study of loading data to snowflake"
authors = ["Toby Devlin <toby@tobydevlin.com>"]
readme = "README.md"
packages = [{include = "westfarmers_case_study"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.21.0"
snowflake-connector-python = {extras = ["pandas"], version = "^2.7.11"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
